Moved settings to server subfolder
Actual addon template contains server subfolder, there was none previously.
